BLOCKCHAIN-523 split proposals into components
